requirements.txt
flask
request
render_template
mysql-connector-python
passlib